requirement.txt
streamlit==1.53.0
langchain==0.2.17
langchain-google-genai==1.0.6
langchain-community==0.2.19
langchain-chroma==1.1.0
langchain-text-splitters==0.2.4
google-generativeai==0.5.4
python-dotenv==1.2.1
gTTS==2.5.4
streamlit_mic_recorder==0.0.8
pypdf==4.1.0
chromadb==1.4.1
pysqlite3-binary